refactored calculateTextSize into seprate script, added it to the global scope of all SchemioScript executions and improved the pie hcart animation
