+++ src/ui/components/editor/items/ItemTemplateFunctions.js
@@ -9,6 +9,7 @@ import EditorEventBus from '../EditorEventBus';
 import { enrichItemWithDefaults } from '../../../scheme/ItemFixer';
 import { createConnectionsFunctions } from '../../../scripting/connections';
 import UserEventBus from '../../../userevents/UserEventBus';
+import { calculateTextSize } from '../../../templater/text';
 
 
 const dummyUserEventBus = new UserEventBus();
@@ -40,25 +41,6 @@ export function createTemplateFunctions(editorId, rootItem, schemeContainer) {
     }
 }
 
-const fontCorrections = {
-    'Lucida Sans Unicode': 1.4,
-    'Lucida Console': 1.47,
-};
-
-export function calculateTextSize(text, font, fontSize) {
-    const canvas = calculateTextSize.canvas || (calculateTextSize.canvas = document.createElement("canvas"));
-    const context = canvas.getContext("2d");
-    context.font = `${fontSize}px ${font}`;
-    const metrics = context.measureText(text);
-    let wk = 1;
-    if (fontCorrections.hasOwnProperty(font)) {
-        wk = fontCorrections[font];
-    }
-    return {
-        w: metrics.width * wk,
-        h: Math.abs(metrics.actualBoundingBoxAscent) + Math.abs(metrics.actualBoundingBoxDescent)
-    };
-}
 
 /**
  * @param {Item} rootItem
